+++ components/radio/legacy-styled.jsx
@@ -1,28 +1,20 @@
 import styled from 'styled-components';
-import { position, space, layout } from 'styled-system';
-import { themeGet } from '@styled-system/theme-get';
-import { Text } from '../Text';
 import { resetStyles } from '../utils';
 
 export const RadioSvg = styled.svg`
-	display: inline;
-	position: relative;
+	position: absolute;
 	width: 14px;
 	height: 14px;
-
-	${position}
-	${space}
-	${layout}
 `;
 
 export const RadioBorder = styled.circle`
 	fill: none;
-	stroke: ${themeGet('colors.radio.border')};
+	stroke: ${props => props.theme.border};
 	stroke-width: 2;
 `;
 
 export const CheckedIndicator = styled.circle`
-	fill: ${themeGet('colors.radio.primary')};
+	fill: ${props => props.theme.primary};
 `;
 
 export const RadioContainer = styled.button`
@@ -44,19 +36,19 @@ export const RadioContainer = styled.button`
 
 		@media (hover: hover) {
 			&:hover ${RadioBorder} {
-				stroke: ${themeGet('colors.radio.primary')};
+				stroke: ${props => props.theme.primary};
 			}
 		}
 		@media (hover: none) {
 			&:active ${RadioBorder} {
-				stroke: ${themeGet('colors.radio.primary')};
+				stroke: ${props => props.theme.primary};
 			}
 		}
 	}
 
 	&:disabled ${RadioBorder} {
-		fill: ${themeGet('colors.radio.disabledBackground')};
-		stroke: ${themeGet('colors.radio.disabledBorder')};
+		fill: ${props => props.theme.disabledBackground};
+		stroke: ${props => props.theme.disabledBorder};
 	}
 
 	&:focus {
@@ -67,11 +59,10 @@ export const RadioContainer = styled.button`
 	}
 `;
 
-/**
- * An optional configuration component that allows styling just the label via Styled System props.
- */
-export const RadioLabel = styled(Text)`
-	margin-left: 6px;
+export const Label = styled.div`
+	margin-left: 22px;
+
+	& + & {
+		margin-left: 6px;
+	}
 `;
-RadioLabel.propTypes = Text.propTypes;
-RadioLabel.childConfigComponent = 'RadioLabel';