+++ index-v6.js
@@ -1,4 +1,8 @@
-export { Modal } from './components/modal';
 export { Button, SegmentedButtonGroup } from './components/button';
-export { usePopover, Popover } from './components/popover-v6';
+export { Checkbox } from './components/check-box';
 export { Dropdown } from './components/dropdown';
+export { Modal } from './components/modal';
+export { usePopover, Popover } from './components/popover-v6';
+export { Radio } from './components/radio';
+export { SimpleToast } from './components/simple-toast';
+export { Slider } from './components/slider';